+++ .env.development
@@ -1,0 +1,1 @@
+#ST_API_SERVER_URL=http://scoretrak.example.com:30080
