--- binding.gyp
@@ -1,8 +1,0 @@
-{
-	"targets": [
-		{
-			"target_name": "binding",
-			"sources": [ "build/Release/binding.node" ]
-		}
-	]
-}
