+++ package.json
@@ -13,16 +13,6 @@
     "@nivo/core": "^0.79.0",
     "@nivo/tooltip": "0.79.0",
     "@tanstack/react-query": "^4.2.1",
-    "@types/file-saver": "^2.0.1",
-    "@types/google-protobuf": "^3.7.4",
-    "@types/jest": "^28.1.6",
-    "@types/material-ui": "^0.21.8",
-    "@types/node": "^16.11.45",
-    "@types/react": "^17.0.47",
-    "@types/react-dom": "^17.0.17",
-    "@types/react-router": "^5.1.18",
-    "@types/react-router-dom": "^5.3.3",
-    "@types/uuid": "^8.3.0",
     "file-saver": "^2.0.2",
     "google-protobuf": "^3.14.0-rc.1",
     "grpc-web": "^1.2.1",
@@ -36,27 +26,13 @@
     "react-hook-form": "^6.11.0",
     "react-json-view": "^1.21.3",
     "react-router-dom": "^6.3.0",
-    "react-scripts": "5.0.1",
-    "react-use": "^17.4.0",
-    "ts-transformer-keys": "^0.4.3",
-    "typescript": "4.7.4",
-    "watch": "^1.0.2",
-    "web-vitals": "^2.1.4"
+    "react-use": "^17.4.0"
   },
   "scripts": {
-    "start": "GENERATE_SOURCEMAP=false react-scripts start",
-    "build": "GENERATE_SOURCEMAP=false react-scripts build",
-    "test": "GENERATE_SOURCEMAP=false react-scripts test",
-    "eject": "GENERATE_SOURCEMAP=false react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ],
-    "ignorePatterns": [
-      "**/*_pb.js"
-    ]
+    "dev": "vite",
+    "dev:clean": "npm run dev -- --force",
+    "build": "tsc && vite build",
+    "serve": "vite preview --port 3000"
   },
   "browserslist": {
     "production": [
@@ -71,13 +47,19 @@
     ]
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0-0",
-    "@babel/plugin-syntax-flow": "^7.14.5",
-    "@babel/plugin-transform-react-jsx": "^7.14.9",
-    "@testing-library/dom": ">=7.21.4",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^12.1.5",
-    "@testing-library/user-event": "^14.2.5",
-    "prettier": "2.7.1"
+    "@types/file-saver": "^2.0.1",
+    "@types/google-protobuf": "^3.7.4",
+    "@types/material-ui": "^0.21.8",
+    "@types/node": "^16.11.45",
+    "@types/react": "^17.0.47",
+    "@types/react-dom": "^17.0.17",
+    "@types/react-router": "^5.1.18",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/uuid": "^8.3.0",
+    "@vitejs/plugin-react": "^2.0.1",
+    "prettier": "2.7.1",
+    "typescript": "^4.6.4",
+    "vite": "^3.0.9",
+    "vite-plugin-env-compatible": "^1.1.1"
   }
 }
