+++ src/layouts/AuthLayout.tsx
@@ -1,11 +1,16 @@
 import { MainNav } from "../components/MainNav";
 import { Outlet } from "react-router-dom";
+import { Box, Container } from "@material-ui/core";
 
 export default function AuthLayout() {
   return (
     <>
       <MainNav />
-      <Outlet />
+      <Box m={2} display={"flex"}>
+        <Container maxWidth={"sm"}>
+          <Outlet />
+        </Container>
+      </Box>
     </>
   );
 }
