+++ src/layouts/DefaultLayout.tsx
@@ -1,12 +1,16 @@
 import { Outlet } from "react-router-dom";
-import BaseLayout from "./BaseLayout";
+import { Box, Container } from "@material-ui/core";
+import { MainNav } from "../components/MainNav";
 
 export default function DefaultLayout() {
   return (
     <>
-      <BaseLayout>
-        <Outlet />
-      </BaseLayout>
+      <MainNav />
+      <Box m={2} display={"flex"}>
+        <Container maxWidth={"xl"}>
+          <Outlet />
+        </Container>
+      </Box>
     </>
   );
 }
