+++ src/layouts/ScoreboardLayout.tsx
@@ -1,28 +1,31 @@
 import { Outlet } from "react-router-dom";
 import { useReport } from "../contexts/ReportContext";
 import React from "react";
-import { Box, CircularProgress } from "@material-ui/core";
-import BaseLayout from "./BaseLayout";
+import { Box, CircularProgress, Container } from "@material-ui/core";
 import PreCompBanner from "../components/PreCompBanner";
+import { MainNav } from "../components/MainNav";
 
 export default function ScoreboardLayout() {
   const report = useReport();
   return (
     <>
-      <BaseLayout>
-        {report && report.Round !== 0 ?
-          <Outlet />
-          :
-          <>
-            <Box mt={2} sx={{textAlign: "center"}}>
-              <CircularProgress />
-              {report?.Round === 0 &&
-                <PreCompBanner />
-              }
-            </Box>
-          </>
-        }
-      </BaseLayout>
+      <MainNav />
+      {report && report.Round !== 0 ?
+        <Box m={2} display={"flex"}>
+          <Container maxWidth={"xl"}>
+            <Outlet />
+          </Container>
+        </Box>
+        :
+        <>
+          <Box mt={2} sx={{textAlign: "center"}}>
+            <CircularProgress />
+            {report?.Round === 0 &&
+              <PreCompBanner />
+            }
+          </Box>
+        </>
+      }
     </>
   );
 }
