--- tslint-imports.json
@@ -1,9 +1,0 @@
-{
-  "extends": ["tslint-etc", "tslint:latest"],
-  "rules": {
-    "no-unused-declaration": true
-  },
-  "linterOptions": {
-    "exclude": ["", "./src/lib/scoretrakapis/**/*.ts"]
-  }
-}
